05-31 13:24 finish generate object code

+++ testmain.cpp
@@ -1,9 +1,9 @@
 #include <iostream>
 
 extern "C" {
-    int64_t do_math(int64_t);
+    int do_math(int);
 }
 
 int main() {
-    std::cout << "domath: " << do_math(1) << std::endl;
+	std::cout << "domath: " << do_math(1) << std::endl;
 }
